Empty database, updating with new

+++ server/database_setup.py
@@ -1,7 +1,7 @@
 
 
 
-from sqlalchemy import Column, ForeignKey, Integer, String
+from sqlalchemy import Column, ForeignKey, Integer, String, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from sqlalchemy import create_engine
@@ -12,10 +12,10 @@ class Bus(Base):
     __tablename__ = 'bus'
 
     id = Column(Integer, primary_key=True)
-    longitude = Column(Double, nullable=False)
-    latitude = Column(Double, nullable=False)
-    next_station = Column(Double, nullable=False)
-    prev_station = Column(Double, nullable=False)
+    longitude = Column(Float, nullable=False)
+    latitude = Column(Float, nullable=False)
+    next_station = Column(Float, nullable=False)
+    prev_station = Column(Float, nullable=False)
     time_to_next_station = Column(String(50), nullable=False)
     capacity = Column(Integer, nullable=False)
 
@@ -37,8 +37,8 @@ class Station(Base):
     __tablename__ = 'station'
 
     id = Column(Integer, primary_key=True)
-    longitude = Column(Double, nullable=False)
-    latitude = Column(Double,  nullable=False)
+    longitude = Column(Float, nullable=False)
+    latitude = Column(Float,  nullable=False)
     name = Column(String(50), nullable=False)
     url_img  = Column(String(150), nullable=False)
